requirements/requirements.txt
amqp==1.4.9
anyjson==0.3.3
argh==0.26.2
asgi-redis==1.4.3
asgiref==1.1.2
async-lru==1.0.1
asyncio==3.4.3
attrs==19.3.0
autobahn==20.1.3
Automat==0.8.0
beautifulsoup4==4.8.2
billiard==3.3.0.23
celery==3.1.20
certifi==2019.11.28
cffi==1.13.2
channel==0.1.0
channels==1.1.5
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
coroutines==0.1.4
cryptography==2.8
daphne==1.4.2
defusedxml==0.6.0
dj-database-url==0.4.1
Django==2.2.9
django-bootstrap3==7.0.1
django-celery==3.3.1
django-crispy-forms==1.8.1
django-google-analytics-app==4.4.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.0.4
djangorestframework==3.11.0
gunicorn==19.6.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Jinja2==2.10
kombu==3.0.37
MarkupSafe==1.1.1
msgpack-python==0.5.6
oauthlib==3.1.0
pathtools==0.1.2
Pillow==7.0.0
psycopg2==2.7.3.2
psycopg2-binary==2.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
pyhocon==0.3.38
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
redis==2.10.6
reportlab==3.2.0
repoze.lru==0.7
requests==2.22.0
requests-oauthlib==1.3.0
Routes==2.4.1
service-identity==18.1.0
six==1.14.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
structlog==17.2.0
tornado==4.5.2
tornado-jinja2==0.2.4
Twisted==19.10.0
txaio==20.1.1
ujson==1.35
urllib3==1.25.8
uvloop==0.13.0
watchdog==0.8.3
whitenoise==3.2.1
zope.interface==4.7.1
